Split _xray module into separate 2d and 3d modules
