+++ scico/linop/xray/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2023-2025 by SCICO Developers
+# Copyright (C) 2023-2026 by SCICO Developers
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SCICO package. Details of the copyright and
 # user license can be found in the 'LICENSE' file distributed with the
@@ -73,7 +73,8 @@ from ._util import (
     rotate_volume,
     volume_alignment_rotation,
 )
-from ._xray import XRayTransform2D, XRayTransform3D
+from ._xray2d import XRayTransform2D
+from ._xray3d import XRayTransform3D
 
 __all__ = [
     "XRayTransform2D",
